requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.2
django-cloudinary-storage==0.3.0
django-daraja==1.3.0
django-environ==0.11.2
django-phonenumber-field==7.3.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
phonenumbers==8.13.31
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
